Test async live processing
This fails currently, and I couldn't get it to work with reasonable effort.

The root cause is that the context manager to make the connection
doesn't work properly. I am not sure how live processing in an async context
should be implemented, since we have to read the data in the proper
sequence from one socket.