Fix flakyness in `ReaderPoolImpl.get_result`
This was mostly (only?) triggered by `MerlinDataSource.stream`, in the
testcase `test_datasource`.

The conditions for this issue are as follows:

- Our `ReaderThread`s have read most of the data from the socket, and we
  are at the end of the stream
- One of the reader threads enters `read_multi_frames`, sees that we are
  done (because the number of frames to read is 0), and returns `True`.
  The thread exits as it doesn't have any work left. At this point, the
  result queue is also still empty.
- Another reader thread is still decoding their stack of frames. When it
  is done, it adds its result to the result queue and also exits.
- From the main thread, `ReaderPoolImpl` checks the status of the reader
  threads. This is where the problem happened: it stopped processing
  even when there were still reader threads running, and possibly items
  in the queue or to be added to the queue in the near future.

This is addressed by properly checking that _all_ threads have exited,
and the output queue is empty.

The test cases have also been adapted to use a larger number of threads
for reading, which increases the likelihood of hitting this problem.