Simulator for Dectris detector
This simulator can replay an acquisition that was recorded with
prototypes/zmq_record.py.

Usage: `libertem-live-dectris-sim <file>`

This is not a complete simulator, but implements just enough
of the API to allow performance
testing and run unit tests of the whole acquisition code.

It reads the metadata from the recording and ensures that
the user settings match the file. For exte and exts triggers
it will assume to be triggered. For inte and ints
it will wait for the 'trigger' command.