Basic X-Spectrum support
Runs with the simulator with default settings.

DectrisLivePartition refactored to FullframeLivePartition for
all detectors that can read a sequence of full frames from the queue.

FIXME probably Merlin can be ported to that as well?

FIXME proper support of basic settings analogous to Dectris
FIXME support for multiple detectors and/or receivers
FIXME support for post decoders?