Implement context manager API for DectrisDetectorConnection
It's really important to properly close the connection at the end. If
this is forgotten, or `close` is not called because of missing exception
handling, this can lead to strange errors. Also re-write all tests cases
to explicitly clean up the connection.

For notebooks, and integration into other software, we do still have to
support the explicit API, as before. This is reflected in the tests,
which mostly use the explicit API.