tweak API: `LiveContext.make_acquisition` is now generic
This means users have to adapt less code when they want to support
multiple detectors. Only the detector connection settings and the
controller are now detector type specific, a normal workflow can be
implemented generically over detector types now.

This includes re-working the `trigger` function to a more complete
concept, that allows to hook into multiple points in the lifecycle
of an acquisition. Right now, `on_ready_for_data` is a direct
replacement for `trigger`.